// src/app/model/userProfile.ts
export interface GitProfile{
    login: string;
    username: string;
    bio: string;
    profile_url: string;
    avatar_url:string;
    email: string;
    company: string;
    location:string;
    followers:number;
    followings: number;
    reposCount:number;
    joined:Date;
    repo_url: string;
}